final commit, edits added

+++ www/employee/delete.php
@@ -8,15 +8,26 @@ class EmployeeDeletePage extends CRUDPage
     {
         parent::prepareData();
 
-        $employeeID = filter_input(INPUT_POST, 'employee_id', FILTER_VALIDATE_INT);
-
-        if(!$employeeID)
+        if (session_status() === PHP_SESSION_NONE) {
+            session_start();
+        }
+        if($_SESSION['admin'])
         {
-            throw new BadRequestException();
+
+            $employeeID = filter_input(INPUT_POST, 'employee_id', FILTER_VALIDATE_INT);
+
+            if(!$employeeID)
+            {
+                throw new BadRequestException();
+            }
+
+            $result = Employee::deleteById($employeeID, $_SESSION['user_id']);
+            $this->redirect(self::ACTION_DELETE, $result);
+        }
+        else{
+            throw new UnauthorizedException();
         }
 
-        $result = Employee::deleteById($employeeID, $_SESSION['user_id']);
-        $this->redirect(self::ACTION_DELETE, $result);
     }
